+++ tests/conftest.py
@@ -1,36 +1,10 @@
 from __future__ import annotations
 
 import functools
-import os
-import tempfile
-from io import StringIO
-from pathlib import Path
-from tempfile import NamedTemporaryFile
-from typing import List, NamedTuple
-from unittest import mock
-
 import pytest
-from snowflake.connector.cursor import SnowflakeCursor
 
 from typer import Typer
 from typer.testing import CliRunner
-from tests.testing_utils.fixtures import *
-
-TEST_DIR = Path(__file__).parent
-
-
-@pytest.fixture(scope="session")
-def test_snowcli_config():
-    test_config = TEST_DIR / "test.toml"
-    with tempfile.NamedTemporaryFile(suffix=".toml", mode="w+") as fh:
-        fh.write(test_config.read_text())
-        fh.flush()
-        yield Path(fh.name)
-
-
-@pytest.fixture(scope="session")
-def test_root_path():
-    return TEST_DIR
 
 
 class SnowCLIRunner(CliRunner):
@@ -49,85 +23,3 @@ class SnowCLIRunner(CliRunner):
             ["--config-file", self.test_snowcli_config, *args[0]],
             **kwargs,
         )
-
-
-@pytest.fixture(scope="function")
-def runner(test_snowcli_config):
-    from snowcli.cli.app import app
-
-    return SnowCLIRunner(app, test_snowcli_config)
-
-
-@pytest.fixture
-def mock_cursor():
-    class MockResultMetadata(NamedTuple):
-        name: str
-
-    class _MockCursor(SnowflakeCursor):
-        def __init__(self, rows: List[tuple], columns: List[str]):
-            super().__init__(mock.Mock())
-            self._rows = rows
-            self._columns = [MockResultMetadata(c) for c in columns]
-
-        def fetchone(self):
-            return self.fetchall()
-
-        def fetchall(self):
-            return self._rows
-
-        @property
-        def description(self):
-            yield from self._columns
-
-        @classmethod
-        def from_input(cls, rows, columns):
-            return cls(rows, columns)
-
-    return _MockCursor.from_input
-
-
-@pytest.fixture()
-def mock_ctx(mock_cursor):
-    class _MockConnectionCtx(mock.MagicMock):
-        def __init__(self, cursor=None, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-            self.queries: List[str] = []
-            self.cs = cursor
-
-        def get_query(self):
-            return "\n".join(self.queries)
-
-        def get_queries(self):
-            return self.queries
-
-        @property
-        def warehouse(self):
-            return "MockWarehouse"
-
-        @property
-        def database(self):
-            return "MockDatabase"
-
-        @property
-        def schema(self):
-            return "MockSchema"
-
-        @property
-        def role(self):
-            return "MockRole"
-
-        @property
-        def host(self):
-            return "account.test.region.aws.snowflakecomputing.com"
-
-        def execute_string(self, query: str):
-            self.queries.append(query)
-            if self.cs:
-                return (self.cs,)
-            else:
-                return (mock_cursor(["row"], []),)
-
-        def execute_stream(self, query: StringIO):
-            return self.execute_string(query.read())
-
-    return lambda cursor=None: _MockConnectionCtx(cursor)
