+++ tests/snowpark/test_compute_pool.py
@@ -1,5 +1,7 @@
 from unittest import mock
 
+from tests.testing_utils.fixtures import *
+
 
 @mock.patch("snowflake.connector.connect")
 def test_create_cp(mock_connector, runner, mock_ctx, snapshot):
