+++ tests/test_snow_connector.py
@@ -5,6 +5,7 @@ from unittest import mock
 import pytest
 
 from snowcli.snow_connector import SnowflakeConnector, SnowflakeCursor
+from tests.testing_utils.fixtures import *
 
 
 # Used as a solution to syrupy having some problems with comparing multilines string
