+++ tests/testing_utils/fixtures.py
@@ -1,15 +1,20 @@
 import os
-from pathlib import Path
-from typing import Generator
-
 import pytest
 import tempfile
 
+from io import StringIO
+from pathlib import Path
+from typing import Generator, List, NamedTuple
+from unittest import mock
+
+from snowflake.connector.cursor import SnowflakeCursor
+from tests.conftest import SnowCLIRunner
 from tests.test_data import test_data
 from tests.testing_utils.files_and_dirs import create_named_file, create_temp_file
 
 
 REQUIREMENTS_SNOWFLAKE = "requirements.snowflake.txt"
+TEST_DIR = Path(__file__).parent.parent
 
 
 @pytest.fixture
@@ -45,6 +50,81 @@ def include_paths_env_variable(other_directory: str) -> Generator:
     os.environ.pop("SNOWCLI_INCLUDE_PATHS")
 
 
+@pytest.fixture()
+def mock_ctx(mock_cursor):
+    class _MockConnectionCtx(mock.MagicMock):
+        def __init__(self, cursor=None, *args, **kwargs):
+            super().__init__(*args, **kwargs)
+            self.queries: List[str] = []
+            self.cs = cursor
+
+        def get_query(self):
+            return "\n".join(self.queries)
+
+        def get_queries(self):
+            return self.queries
+
+        @property
+        def warehouse(self):
+            return "MockWarehouse"
+
+        @property
+        def database(self):
+            return "MockDatabase"
+
+        @property
+        def schema(self):
+            return "MockSchema"
+
+        @property
+        def role(self):
+            return "MockRole"
+
+        @property
+        def host(self):
+            return "account.test.region.aws.snowflakecomputing.com"
+
+        def execute_string(self, query: str):
+            self.queries.append(query)
+            if self.cs:
+                return (self.cs,)
+            else:
+                return (mock_cursor(["row"], []),)
+
+        def execute_stream(self, query: StringIO):
+            return self.execute_string(query.read())
+
+    return lambda cursor=None: _MockConnectionCtx(cursor)
+
+
+@pytest.fixture
+def mock_cursor():
+    class MockResultMetadata(NamedTuple):
+        name: str
+
+    class _MockCursor(SnowflakeCursor):
+        def __init__(self, rows: List[tuple], columns: List[str]):
+            super().__init__(mock.Mock())
+            self._rows = rows
+            self._columns = [MockResultMetadata(c) for c in columns]
+
+        def fetchone(self):
+            return self.fetchall()
+
+        def fetchall(self):
+            return self._rows
+
+        @property
+        def description(self):
+            yield from self._columns
+
+        @classmethod
+        def from_input(cls, rows, columns):
+            return cls(rows, columns)
+
+    return _MockCursor.from_input
+
+
 @pytest.fixture
 def other_directory() -> Generator:
     tmp_dir = tempfile.TemporaryDirectory()
@@ -66,6 +146,13 @@ def package_file():
         yield create_named_file("app.zip", tmp, [])
 
 
+@pytest.fixture(scope="function")
+def runner(test_snowcli_config):
+    from snowcli.cli.app import app
+
+    return SnowCLIRunner(app, test_snowcli_config)
+
+
 @pytest.fixture
 def temp_dir():
     initial_dir = os.getcwd()
@@ -85,6 +172,20 @@ def temp_directory_for_app_zip(temp_dir) -> Generator:
 @pytest.fixture
 def temp_file_in_other_directory(other_directory: str) -> Generator:
     yield create_temp_file(".txt", other_directory, [])
+
+
+@pytest.fixture(scope="session")
+def test_snowcli_config():
+    test_config = TEST_DIR / "test.toml"
+    with tempfile.NamedTemporaryFile(suffix=".toml", mode="w+") as fh:
+        fh.write(test_config.read_text())
+        fh.flush()
+        yield Path(fh.name)
+
+
+@pytest.fixture(scope="session")
+def test_root_path():
+    return TEST_DIR
 
 
 @pytest.fixture
